refactored `common` and `configparser`
optimized imports and dependencies in `common` and `configparser`
in preparation to splitting it out of zerocloud
